requirements.txt
numpy>=1.21.0
pandas>=1.3.0
biopython>=1.79
PyYAML>=6.0
tqdm>=4.62.0
ete3>=3.1.2
dendropy>=4.5.2
scipy>=1.7.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
seaborn>=0.11.0
